requirements.txt
feedparser
google-generativeai
pandas
pymongo
certifi
schedule